update patitents page

+++ test/test/Patients_Abrar.aspx.cs
@@ -17,6 +17,7 @@ namespace test
         protected void Page_Load(object sender, EventArgs e)
         {
 
+            // this code for to load the dropdown list from table and display
 
             //string query = "Select department from tblpatients ; ";
 
@@ -47,21 +48,27 @@ namespace test
 
             string query = "Select * from tblpatients Where OHIP = @PatientId  ; ";
 
-            using (SqlConnection conn = new SqlConnection(cs))
-            {
-                SqlCommand cmd = new SqlCommand(query, conn);
-                cmd.Parameters.AddWithValue("@PatientId", patientId);
-                conn.Open();
+            //string query = "Select OHIP,first_name,last_name,phone,date_in,department from tblpatients Where OHIP = @PatientId  ; ";
+            //this query for databind
 
-                SqlDataReader reader = cmd.ExecuteReader();
+            using (SqlConnection conn = new SqlConnection(cs))  //create sql connection
+            {
+                SqlCommand cmd = new SqlCommand(query, conn);   //create sqlcommand
+                cmd.Parameters.AddWithValue("@PatientId", patientId);
+                conn.Open();  // open connection
+
+                SqlDataReader reader = cmd.ExecuteReader(); // SqlDataReader is return type of ExecuteReader
+                //we need object of this class that can fetch the data from the database server.
+
                 if (reader.Read())
                 {
                     txtFirstName.Text = reader["first_name"].ToString();
                     txtLastName.Text = reader["last_name"].ToString();
                     txtPhone.Text = reader["phone"].ToString();
                     txtDateIn.Text = reader["date_in"].ToString();
-                    //txtDateOut.Text = reader["date_out"].ToString();
-                    ddlDepartment.SelectedItem.Text = reader["department"].ToString();
+
+                    //ddlDepartment.SelectedItem.Text = reader["department"].ToString();
+                    txtDepartment.Text = reader["department"].ToString();
                     lblMessage.Text = "Patient with ID = " + patientId + " was found";
                     lblMessage.ForeColor = Color.Green;
 
@@ -75,8 +82,8 @@ namespace test
                     txtLastName.Text = "";
                     txtPhone.Text = "";
                     txtDateIn.Text = "";
-
-                    ddlDepartment.SelectedIndex = -1;
+                    txtDepartment.Text = "";
+                    //ddlDepartment.SelectedIndex = -1;
                     lblMessage2.Text = "";
                 }
             }
@@ -89,14 +96,22 @@ namespace test
             string LastName = txtLastName.Text.Trim();
             string Phone = txtPhone.Text.Trim();
             //DateTime Date_In = Convert.ToDateTime(txtDateIn.Text.Trim());
+
+            string Department = txtDepartment.Text.Trim();
+
             string Date_In = Convert.ToString(txtDateIn.Text.Trim());
+
             //DateTime Date_Out = Convert.ToDateTime(txtDateOut.Text.Trim());
-            string Department = ddlDepartment.SelectedItem.Text;
+            //string Department = ddlDepartment.SelectedItem.Text;
+
 
             string query = "Insert Into tblPatients(OHIP,first_name,last_name,phone,date_in,department)values(@patientId,@FirstName,@LastName,@Phone,@Date_In,@Department);";
-            using (SqlConnection conn = new SqlConnection(cs))
+            //string query1 = "select department from tblPatients ;";
+
+            using (SqlConnection conn = new SqlConnection(cs)) //create sql connection
             {
-                SqlCommand cmd = new SqlCommand(query, conn);
+                SqlCommand cmd = new SqlCommand(query, conn);  //create sqlcommand
+
                 cmd.Parameters.AddWithValue("@patientId", patientId);
                 cmd.Parameters.AddWithValue("@FirstName", FirstName);
                 cmd.Parameters.AddWithValue("@LastName", LastName);
@@ -104,6 +119,10 @@ namespace test
                 cmd.Parameters.AddWithValue("@Date_In", Date_In);
                 //cmd.Parameters.AddWithValue("@Date_Out", Date_Out);
                 cmd.Parameters.AddWithValue("@Department", Department);
+
+                //SqlCommand cmd1 = new SqlCommand(query1, conn);
+                //cmd1.Parameters.AddWithValue("@Department", Department);
+                //DataTable dtTable = new DataTable();
 
 
                 conn.Open();
@@ -120,7 +139,8 @@ namespace test
                     txtPhone.Text = "";
                     txtDateIn.Text = "";
                     lblMessage2.Text = "";
-                    ddlDepartment.SelectedItem.Text = "";
+                    //ddlDepartment.SelectedItem.Text = "";
+                    txtDepartment.Text = "";
                 }
                 else
                 {
@@ -131,9 +151,19 @@ namespace test
                     txtPhone.Text = "";
                     txtDateIn.Text = "";
                     lblMessage2.Text = "";
-                    ddlDepartment.SelectedIndex = -1;
-                }
+                    txtDepartment.Text = "";
+                    //ddlDepartment.SelectedIndex = -1;
 
+                }
+                //SqlDataReader reader1 = cmd1.ExecuteReader();
+                //dtTable.Load(reader1);
+                //reader1.Close();
+                //ddlDepartment.DataSource = dtTable;
+                //ddlDepartment.DataTextField = "department";
+                //ddlDepartment.DataValueField = "department";
+                //ddlDepartment.DataBind();
+                //ddlDepartment.Items.Insert(0, new ListItem("--Select Department--"));
+                //conn.Close();
 
             }
 
@@ -142,17 +172,17 @@ namespace test
 
         protected void txtUpdate_Click(object sender, EventArgs e)
         {
-            if (txtPatientId.Text != "" && txtFirstName.Text != "" && txtLastName.Text != "" && txtPhone.Text != "" && txtDateIn.Text != "")
+            if (txtPatientId.Text != "" && txtFirstName.Text != "" && txtLastName.Text != "" && txtPhone.Text != "" && txtDateIn.Text != "" && txtDepartment.Text != "")
             {
                 string patientId = Convert.ToString(txtPatientId.Text);
                 string FirstName = txtFirstName.Text.Trim();
                 string LastName = txtLastName.Text.Trim();
                 string Phone = txtPhone.Text.Trim();
                 string Date_In = Convert.ToString(txtDateIn.Text.Trim());
-                //DateTime Date_In = Convert.ToDateTime(txtDateIn.Text.Trim());
-                //DateTime Date_Out = Convert.ToDateTime(txtDateOut.Text.Trim());
-                //string Department = ddlDepartment.SelectedIndex.ToString();
-                string Department = ddlDepartment.SelectedItem.Text;
+
+                string Department = txtDepartment.Text.Trim();
+                //string Department = ddlDepartment.SelectedItem.Text;
+
 
                 string query = "Update tblPatients set first_name = @FirstName,last_name = @LastName,phone = @Phone,date_in = @Date_In,department = @Department Where OHIP = @patientId ;";
                 using (SqlConnection conn = new SqlConnection(cs))
@@ -179,7 +209,8 @@ namespace test
                         txtLastName.Text = "";
                         txtPhone.Text = "";
                         txtDateIn.Text = "";
-                        ddlDepartment.SelectedItem.Text = "";
+                        txtDepartment.Text = "";
+                        //ddlDepartment.SelectedItem.Text = "";
 
                     }
                     else
@@ -190,8 +221,8 @@ namespace test
                         txtLastName.Text = "";
                         txtPhone.Text = "";
                         txtDateIn.Text = "";
-
-                        ddlDepartment.SelectedIndex = -1;
+                        txtDepartment.Text = "";
+                        //ddlDepartment.SelectedIndex = -1;
                     }
 
 
@@ -205,17 +236,17 @@ namespace test
         }
         protected void txtDelete_Click(object sender, EventArgs e)
         {
-            if (txtPatientId.Text != "" && txtFirstName.Text != "" && txtLastName.Text != "" && txtPhone.Text != "" && txtDateIn.Text != "")
+            if (txtPatientId.Text != "" && txtFirstName.Text != "" && txtLastName.Text != "" && txtPhone.Text != "" && txtDateIn.Text != "" && txtDepartment.Text != "")
             {
                 string patientId = Convert.ToString(txtPatientId.Text);
                 string FirstName = txtFirstName.Text.Trim();
                 string LastName = txtLastName.Text.Trim();
                 string Phone = txtPhone.Text.Trim();
                 string Date_In = Convert.ToString(txtDateIn.Text.Trim());
-                //DateTime Date_In = Convert.ToDateTime(txtDateIn.Text.Trim());
-                //DateTime Date_Out = Convert.ToDateTime(txtDateOut.Text.Trim());
-                //string Department = ddlDepartment.SelectedIndex.ToString();
-                string Department = ddlDepartment.SelectedItem.Text;
+
+                string Department = txtDepartment.Text.Trim();
+
+                //string Department = ddlDepartment.SelectedItem.Text;
 
                 string query = "Delete from tblPatients Where OHIP = @patientId and first_name = @FirstName and last_name = @LastName and phone = @Phone and date_in = @Date_In and department = @Department ;";
                 using (SqlConnection conn = new SqlConnection(cs))
@@ -242,7 +273,8 @@ namespace test
                         txtLastName.Text = "";
                         txtPhone.Text = "";
                         txtDateIn.Text = "";
-                        ddlDepartment.SelectedItem.Text = "";
+                        txtDepartment.Text = "";
+                        //ddlDepartment.SelectedItem.Text = "";
                         lblMessage2.Text = "";
                     }
                     else
@@ -254,7 +286,8 @@ namespace test
                         txtPhone.Text = "";
                         txtDateIn.Text = "";
                         lblMessage2.Text = "";
-                        ddlDepartment.SelectedIndex = -1;
+                        txtDepartment.Text = "";
+                        //ddlDepartment.SelectedIndex = -1;
                     }
 
 
@@ -277,16 +310,17 @@ namespace test
             SqlConnection conn = new SqlConnection(cs);
             SqlCommand cmd = new SqlCommand(query, conn);
 
-            conn.Open();
+            conn.Open();  //open the connection 
 
-            SqlDataReader reader = cmd.ExecuteReader();
+            SqlDataReader reader = cmd.ExecuteReader(); // SqlDataReader is return type of ExecuteReader
+                                                        //we need object of this class that can fetch the data from the database server.
 
             GridView1.DataSource = reader;
             GridView1.DataBind();
 
-            conn.Close();
+            conn.Close(); //close the connection 
             lblMessage.Text = "";
-            lblMessage2.Text = "";
+
         }
 
         protected void GridView1_SelectedIndexChanged(object sender, EventArgs e)
@@ -301,16 +335,14 @@ namespace test
             txtLastName.Text = "";
             txtPhone.Text = "";
             txtDateIn.Text = "";
+            txtDepartment.Text = "";
 
             //ddlDepartment.SelectedIndex = -1;
             //ddlDepartment.SelectedIndex = ddlDepartment.Items.IndexOf(ddlDepartment.Items.FindByValue("Select"));
             //ddlDepartment.ClearSelection();
         }
 
-        protected void ddlDepartment_SelectedIndexChanged(object sender, EventArgs e)
-        {
 
-        }
     }
 
 }
